Factor Game query into its own hook
This hook is responsible for fetching the object and also parsing it.
Moving the parsing logic into the hook also makes it easier to deal
with pending and error states, and later, dependenct queries, which we
will need to support querying whether the game has ended or not using
a dev inspect query.